:sparkles: feat(product): 상품 price int로 변경후 상품 수량 추가 및 결제 시 수량 감소 구현 (#24)
* :sparkles: feat(product): 상품 price int로 변경후 상품 수량 추가 및 결제 시 수량 감소 구현

* :recycle: fix(product): 주문 테이블 이름 수정

* :recycle: fix(product): 주문 로직 수정

+++ src/main/java/com/kuku9/goods/domain/seller/service/SellerServiceImpl.java
@@ -99,9 +99,9 @@ public class SellerServiceImpl implements SellerService {
         }
 
         List<SellingProductResponse> responseDtoList = new ArrayList<>();
-        long totalPrice = 0L;
+        int totalPrice = 0;
         for (OrderProduct orderProduct : orderProductList) {
-            long productTotalPrice =
+            int productTotalPrice =
                 orderProduct.getProduct().getPrice() * orderProduct.getQuantity();
             responseDtoList.add(
                 new SellingProductResponse(
@@ -127,10 +127,10 @@ public class SellerServiceImpl implements SellerService {
             }
         }
 
-        long totalPrice = 0L;
-        long statisticsPrice = 0L;
+        int totalPrice = 0;
+        int statisticsPrice = 0;
         for (OrderProduct orderProduct : orderProductList) {
-            long productTotalPrice =
+            int productTotalPrice =
                 (orderProduct.getProduct().getPrice() * orderProduct.getQuantity());
             totalPrice += productTotalPrice;
         }
